:truck: separate types & index for export on avatargroup
