+++ packages/react/src/components/AvatarGroup/AvatarGroup.types.ts
@@ -1,0 +1,8 @@
+import type { AvatarProps } from '../Avatar/Avatar.types';
+
+export type AvatarGroupProps = {
+  avatars: AvatarProps[];
+  max?: number;
+  spacing?: 'small' | 'medium' | 'large';
+  size?: 'small' | 'medium' | 'large';
+};
