+++ packages/react/src/components/AvatarGroup/index.ts
@@ -1,1 +1,2 @@
 export { default as AvatarGroup } from './avatarGroup';
+export * from './AvatarGroup.types';
